refactor: moved useChainLinksGivenAddress into hooks/chainlinks
